requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.10
numpy==1.24.3
opencv-python==4.9.0.80
PyPDF2==3.0.1
PyMuPDF==1.23.26